CLI and client improvements. (#120)

+++ resolvers/dotnet/Azure.IoT.DeviceModelsRepository.CLI/ModelImporter.cs
@@ -1,4 +1,5 @@
 ﻿using Azure.IoT.DeviceModelsRepository.Resolver;
+using System;
 using System.IO;
 using System.Text;
 
@@ -8,8 +9,18 @@ namespace Azure.IoT.DeviceModelsRepository.CLI
     {
         public static void Import(string modelContent, DirectoryInfo repository)
         {
-            string rootId = new ModelQuery(modelContent).GetId();
+            string rootId = Parsing.GetRootId(modelContent);
             string createPath = DtmiConventions.DtmiToQualifiedPath(rootId, repository.FullName);
+
+            Outputs.WriteOut($"- Importing model \"{rootId}\"...");
+            if (File.Exists(createPath))
+            {
+                Outputs.WriteOut(
+                    $"Skipping \"{rootId}\". Model file already exists in repository.",
+                    ConsoleColor.DarkCyan);
+                return;
+            }
+
             (new FileInfo(createPath)).Directory.Create();
             File.WriteAllText(createPath, modelContent, Encoding.UTF8);
         }
